# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-Ebplex ",
        "description": "Swagger OpenApi description",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/client/{id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Get information",
                "description": "Returns data",
                "operationId": "App\\Http\\Controllers\\ClientController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": " Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Resource Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successful operation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Client"
                ],
                "summary": "Update existing",
                "description": "Returns updated clientes",
                "operationId": "profileUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "types_id",
                                    "identification",
                                    "lastname",
                                    "name",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "types_id": {
                                        "type": "numeric"
                                    },
                                    "identification": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user2@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, client does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Element obtained successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Delete existing ",
                "description": "Deletes a record and returns no content",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This element cannot be removed, because it has dependent relationships"
                    }
                }
            }
        },
...
        "/api/contract/{id}": {
            "get": {
                "tags": [
                    "Contract"
                ],
                "summary": "Get information",
                "description": "Returns data",
                "operationId": "App\\Http\\Controllers\\ContractController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contract id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": " Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Resource Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successful operation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contract"
                ],
                "summary": "Update existing",
                "description": "Returns updated contract",
                "operationId": "profileUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clients_id",
                                    "n_contract",
                                    "date"
                                ],
                                "properties": {
                                    "clients_id": {
                                        "type": "numeric"
                                    },
                                    "n_contract": {
                                        "type": "numeric"
                                    },
                                    "date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Element obtained successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contract"
                ],
                "summary": "Delete existing ",
                "description": "Deletes a record and returns no content",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "contract id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed successfully"
                    }
                }
            }
        },
...
        "/api/datacompany/{id}": {
            "get": {
                "tags": [
                    "Data Company"
                ],
                "summary": "Get information",
                "description": "Returns data",
                "operationId": "App\\Http\\Controllers\\DataCompanyController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "data company ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": " Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Resource Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successful operation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Data Company"
                ],
                "summary": "Update existing",
                "description": "Returns updated data company",
                "operationId": "profileUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "datacompany id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "owner_clients_id",
                                    "bussiness_name",
                                    "legal_representative"
                                ],
                                "properties": {
                                    "owner_clients_id": {
                                        "type": "numeric"
                                    },
                                    "bussiness_name": {
                                        "type": "string"
                                    },
                                    "legal_representative": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Element obtained successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Data Company"
                ],
                "summary": "Delete existing ",
                "description": "Deletes a record and returns no content",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Data Company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed successfully"
                    }
                }
            }
        },
...
        "/api/sex_user_data": {
            "get": {
                "tags": [
                    "Sex User Data "
                ],
                "summary": "Get list of Sex User Data ",
                "description": "Returns list ",
                "operationId": "App\\Http\\Controllers\\SexUserDataController::index",
                "responses": {
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successful operation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sex User Data "
                ],
                "summary": "Store new Sex User Data",
                "description": "Returns  data",
                "operationId": "App\\Http\\Controllers\\SexUserDataController::store",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Bad Request, Unprocessable Entity "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Element obtained successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "successful operation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}